+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Library] `BulletList` component tokens (tokens library v0.9.0)
 - [Library] `Chip` component tokens (tokens library v0.9.0)
+- [Library] Use of color scheme to apply elevation effects ([#146](https://github.com/Orange-OpenSource/ouds-ios/issues/146)) 
 - [Library] Remove UI test cases on colored surface for radio button and checkbox ([#523](https://github.com/Orange-OpenSource/ouds-ios/issues/523)) 
 - [Library] Rafactor the interal interation state used by radio and checkbox components ([#560](https://github.com/Orange-OpenSource/ouds-ios/issues/560)) 
 - [Library] Font family, font weight semantic tokens (tokens library v0.8.0) ([#529](https://github.com/Orange-OpenSource/ouds-ios/issues/529))